requirements.txt
fastapi
uvicorn[standard]
gradio>=4.0.0

transformers>=4.45.0
accelerate
huggingface-hub

pillow
numpy
python-multipart
pydantic>=2
starlette
requests
tqdm
num2words
